Some fixes to the test_expired_metadata
The test needs some modifications in order to do what is supposed to.

First, we need a current timestamp/snapshot version 1 which is expired.
Second, we need a second metadata version 2 which is valid with a
future expiration date.
I made changes to the test to make sure we have the first and second
sets of metadata are available and downloaded in the correct order.

Finally, I added a check, in the end, to verify that the stored metadata
files after the second update are version 2 with a future expiry date.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>