Add targets to the test_expired_metadata
Targets is a special case regarding the test_expired_metadata.
As opposed to timestamp and snapshot if a local targets file has
expired then we don't use any information from it to verify the next
target's version.

In conclusion: if a local targets has expired the update process won't
stop, but the local targets won't be verified and the update will try
to fetch and verify the next version without using any information from
the old expired targets file.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>